Refactor(fe): 컴포넌트 및 타입 재배치
