--- frontend/grabtable/src/app/types/cartResponse.ts
@@ -1,7 +1,0 @@
-export type CartResponse = {
-  id: number
-  menuName: string
-  price: number
-  quantity: number
-  totalPrice: number
-}
