--- frontend/grabtable/src/app/types/orderResponse.ts
@@ -1,9 +1,0 @@
-import { CartResponse } from './cartResponse'
-
-export type OrderResponse = {
-  id: number
-  userId: number
-  paidAmount: number
-  carts: CartResponse[]
-  status: string
-}
