--- frontend/grabtable/src/app/types/postOrderResponse.ts
@@ -1,8 +1,0 @@
-import { Cart } from './cart'
-
-export type PostOrderResponse = {
-  id: number
-  userId: number
-  carts: Cart[]
-  status: string
-}
