--- frontend/grabtable/src/app/types/sharedOrderResponse.ts
@@ -1,10 +1,0 @@
-import { CartResponse } from './cartResponse'
-import { OrderResponse } from './orderResponse'
-
-export type SharedOrderResponse = {
-  id: number
-  carts: CartResponse[]
-  orders: OrderResponse[]
-  totalAmount: number
-  leftAmount: number
-}
