--- frontend/grabtable/src/app/types/storeResponse.ts
@@ -1,8 +1,0 @@
-export type StoreResponse = {
-  address: string
-  averageRating: number
-  category: string
-  id: number
-  storeName: string
-  storePictureUrl: string
-}
