--- frontend/grabtable/src/app/types/userCartsInfoResponse.ts
@@ -1,8 +1,0 @@
-import { CartResponse } from './cartResponse'
-
-export type UserCartsInfoResponse = {
-  id: number
-  username: string
-  profileImageUrl: string
-  currentCarts: CartResponse[]
-}
